requirements.txt
python-binance==1.0.19
pandas==2.1.4
numpy==1.26.2
python-telegram-bot==20.7
aiohttp==3.9.1
asyncio==3.4.3
ta==0.11.0
python-dotenv==1.0.0
newsapi-python==0.2.7
tweepy==4.14.0
beautifulsoup4==4.12.2
requests==2.31.0 